BLUEDOC-94 - Create work type: Umrdr / DataSet
