+++ .gitignore
@@ -17,6 +17,9 @@
 /tmp/*
 !/log/.keep
 !/tmp/.keep
+fits.log
 
 # Ignore Byebug command history file.
 .byebug_history
+
+.DS_Store
