+++ app/actors/hyrax/actors/data_set_actor.rb
@@ -3,6 +3,7 @@
 module Hyrax
   module Actors
     class DataSetActor < Hyrax::Actors::BaseActor
+      # nothing added from Hyrax Umrdr v1
     end
   end
 end
