--- app/actors/hyrax/actors/umrdr_work_actor.rb
@@ -1,9 +1,0 @@
-# Generated via
-#  `rails generate hyrax:work UmrdrWork`
-module Hyrax
-  module Actors
-    class UmrdrWorkActor < Hyrax::Actors::BaseActor
-      # nothing added from Hyrax Umrdr v1
-    end
-  end
-end
