--- app/controllers/concerns/umrdr/file_set_behavior.rb
@@ -1,10 +1,0 @@
-module Umrdr
-  module FileSetBehavior
-    extend ActiveSupport::Concern
-
-    def update_parent()
-      parent.total_file_size_add_file_set!( self ) unless parent.nil?
-    end
-
-  end
-end
