--- app/controllers/concerns/umrdr/works_controller_behavior.rb
@@ -1,18 +1,0 @@
-module Umrdr
-  module WorksControllerBehavior
-    extend ActiveSupport::Concern
-    #in umrdr
-    #include Hyrax::Controller
-    include Deepblue::WorksControllerBehavior
-
-    # override curation concerns, add form fields values
-    def build_form
-      super
-      # Set up the multiple parameters for the date coverage attribute in the form
-      cov_date = Date.edtf(@form.date_coverage.first)
-      cov_params = Umrdr::DateCoverageService.interval_to_params cov_date
-      @form.merge_date_coverage_attributes! cov_params
-    end
-
-  end
-end
