--- app/views/hyrax/umrdr_works/_umrdr_work.html.erb
@@ -1,2 +1,0 @@
-<%# This is a search result view %>
-<%= render 'catalog/document', document: umrdr_work, document_counter: umrdr_work_counter  %>
