+++ config/initializers/hyrax.rb
@@ -1,8 +1,6 @@
 Hyrax.config do |config|
   # Injected via `rails g hyrax:work DataSet`
   config.register_curation_concern :data_set
-  # Injected via `rails g hyrax:work UmrdrWork`
-  config.register_curation_concern :umrdr_work
   # Register roles that are expected by your implementation.
   # @see Hyrax::RoleRegistry for additional details.
   # @note there are magical roles as defined in Hyrax::RoleRegistry::MAGIC_ROLES
