--- config/locales/umrdr_work.en.yml
@@ -1,8 +1,0 @@
-en:
-  hyrax:
-    icons:
-      umrdr_work:     'fa fa-file-text-o'
-    select_type:
-      umrdr_work:
-        description:        "Umrdr work works"
-        name:               "Umrdr Work"
