--- config/locales/umrdr_work.es.yml
@@ -1,10 +1,0 @@
-es:
-  hyrax:
-    icons:
-      umrdr_work:     'fa fa-file-text-o'
-    select_type:
-      umrdr_work:
-        # TODO: translate `human_name` into Spanish
-        description:        "Umrdr work trabajos"
-        name:               "Umrdr Work"
-        # TODO: translate `human_name` into Spanish
