--- config/locales/umrdr_work.fr.yml
@@ -1,8 +1,0 @@
-fr:
-  hyrax:
-    icons:
-      umrdr_work:     'fa fa-file-text-o'
-    select_type:
-      umrdr_work:
-        description:        "Umrdr work œuvres"
-        name:               "Umrdr Work"
