--- config/locales/umrdr_work.pt-BR.yml
@@ -1,8 +1,0 @@
-pt-BR:
-  hyrax:
-    icons:
-      umrdr_work:     'fa fa-file-text-o'
-    select_type:
-      umrdr_work:
-        description:        "Umrdr work trabalhos"
-        name:               "Umrdr Work"
