--- config/locales/umrdr_work.zh.yml
@@ -1,10 +1,0 @@
-zh:
-  hyrax:
-    icons:
-      umrdr_work:     'fa fa-file-text-o'
-    select_type:
-      umrdr_work:
-        # TODO: translate `human_name` into Chinese
-        description:        "Umrdr work 作品"
-        name:               "Umrdr Work"
-        # TODO: translate `human_name` into Chinese
