--- spec/controllers/hyrax/umrdr_works_controller_spec.rb
@@ -1,9 +1,0 @@
-# Generated via
-#  `rails generate hyrax:work UmrdrWork`
-require 'rails_helper'
-
-RSpec.describe Hyrax::UmrdrWorksController do
-  it "has tests" do
-    skip "Add your tests here"
-  end
-end
