--- spec/forms/hyrax/umrdr_work_form_spec.rb
@@ -1,9 +1,0 @@
-# Generated via
-#  `rails generate hyrax:work UmrdrWork`
-require 'rails_helper'
-
-RSpec.describe Hyrax::UmrdrWorkForm do
-  it "has tests" do
-    skip "Add your tests here"
-  end
-end
